@@ -1,49 +1,78 @@
 module github.com/rusq/slackdump/v2
 
-go 1.18
+go 1.21.4
 
 require (
-	github.com/AlecAivazis/survey/v2 v2.3.5
 	github.com/MercuryEngineering/CookieMonster v0.0.0-20180304172713-1584578b3403
+	github.com/charmbracelet/huh v0.2.3
 	github.com/denisbrodbeck/machineid v1.0.1
-	github.com/fatih/color v1.13.0
-	github.com/gdamore/tcell/v2 v2.5.2
-	github.com/golang/mock v1.6.0
-	github.com/joho/godotenv v1.4.0
-	github.com/pkg/browser v0.0.0-20210911075715-681adbf594b8
-	github.com/playwright-community/playwright-go v0.2000.1
-	github.com/rivo/tview v0.0.0-20220728094620-c6cff75ed57b
-	github.com/rusq/dlog v1.3.3
+	github.com/fatih/color v1.15.0
+	github.com/gdamore/tcell/v2 v2.7.0
+	github.com/joho/godotenv v1.5.1
+	github.com/pkg/browser v0.0.0-20240102092130-5ac0b6a4141c
+	github.com/playwright-community/playwright-go v0.4001.0
+	github.com/rivo/tview v0.0.0-20240101144852-b3bd1aa5e9f2
+	github.com/rusq/chttp v1.0.2
+	github.com/rusq/dlog v1.4.0
 	github.com/rusq/osenv/v2 v2.0.1
-	github.com/rusq/secure v0.0.3
+	github.com/rusq/secure v0.0.4
+	github.com/rusq/slackauth v0.0.7
 	github.com/rusq/tracer v1.0.1
-	github.com/schollz/progressbar/v3 v3.8.6
-	github.com/slack-go/slack v0.11.0
-	github.com/stretchr/testify v1.7.1
-	golang.org/x/time v0.0.0-20220722155302-e5dcc9cfc0b9
+	github.com/schollz/progressbar/v3 v3.13.0
+	github.com/slack-go/slack v0.12.1
+	github.com/stretchr/testify v1.8.4
+	go.uber.org/mock v0.4.0
+	golang.org/x/sync v0.4.0
+	golang.org/x/term v0.15.0
+	golang.org/x/time v0.3.0
 )
 
 require (
+	github.com/alecthomas/chroma v0.10.0 // indirect
+	github.com/atotto/clipboard v0.1.4 // indirect
+	github.com/aymanbagabas/go-osc52/v2 v2.0.1 // indirect
+	github.com/aymerick/douceur v0.2.0 // indirect
+	github.com/catppuccin/go v0.2.0 // indirect
+	github.com/charmbracelet/bubbles v0.17.1 // indirect
+	github.com/charmbracelet/bubbletea v0.25.0 // indirect
+	github.com/charmbracelet/glamour v0.6.0 // indirect
+	github.com/charmbracelet/lipgloss v0.9.1 // indirect
+	github.com/containerd/console v1.0.4-0.20230313162750-1ae8d489ac81 // indirect
 	github.com/danwakefield/fnmatch v0.0.0-20160403171240-cbb64ac3d964 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/dlclark/regexp2 v1.10.0 // indirect
 	github.com/gdamore/encoding v1.0.0 // indirect
+	github.com/go-jose/go-jose/v3 v3.0.1 // indirect
+	github.com/go-rod/rod v0.114.5 // indirect
 	github.com/go-stack/stack v1.8.1 // indirect
-	github.com/gorilla/websocket v1.5.0 // indirect
-	github.com/kballard/go-shellquote v0.0.0-20180428030007-95032a82bc51 // indirect
+	github.com/gorilla/css v1.0.0 // indirect
+	github.com/gorilla/websocket v1.5.1 // indirect
+	github.com/kr/text v0.2.0 // indirect
 	github.com/lucasb-eyer/go-colorful v1.2.0 // indirect
-	github.com/mattn/go-colorable v0.1.12 // indirect
-	github.com/mattn/go-isatty v0.0.14 // indirect
-	github.com/mattn/go-runewidth v0.0.13 // indirect
-	github.com/mgutz/ansi v0.0.0-20200706080929-d51e80ef957d // indirect
+	github.com/mattn/go-colorable v0.1.13 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
+	github.com/mattn/go-localereader v0.0.1 // indirect
+	github.com/mattn/go-runewidth v0.0.15 // indirect
+	github.com/microcosm-cc/bluemonday v1.0.25 // indirect
 	github.com/mitchellh/colorstring v0.0.0-20190213212951-d06e56a500db // indirect
+	github.com/muesli/ansi v0.0.0-20230316100256-276c6243b2f6 // indirect
+	github.com/muesli/cancelreader v0.2.2 // indirect
+	github.com/muesli/reflow v0.3.0 // indirect
+	github.com/muesli/termenv v0.15.2 // indirect
+	github.com/olekukonko/tablewriter v0.0.5 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
-	github.com/rivo/uniseg v0.3.1 // indirect
-	golang.org/x/crypto v0.0.0-20220722155217-630584e8d5aa // indirect
-	golang.org/x/sys v0.0.0-20220803195053-6e608f9ce704 // indirect
-	golang.org/x/term v0.0.0-20220722155259-a9ba230a4035 // indirect
-	golang.org/x/text v0.3.7 // indirect
-	gopkg.in/square/go-jose.v2 v2.6.0 // indirect
-	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+	github.com/rivo/uniseg v0.4.4 // indirect
+	github.com/ysmood/fetchup v0.2.4 // indirect
+	github.com/ysmood/goob v0.4.0 // indirect
+	github.com/ysmood/got v0.38.2 // indirect
+	github.com/ysmood/gson v0.7.3 // indirect
+	github.com/ysmood/leakless v0.8.0 // indirect
+	github.com/yuin/goldmark v1.6.0 // indirect
+	github.com/yuin/goldmark-emoji v1.0.2 // indirect
+	go.uber.org/multierr v1.11.0 // indirect
+	golang.org/x/crypto v0.17.0 // indirect
+	golang.org/x/net v0.17.0 // indirect
+	golang.org/x/sys v0.15.0 // indirect
+	golang.org/x/text v0.14.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
-
-replace github.com/slack-go/slack => github.com/rusq/slack v0.11.100
