@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: slackdump.go
-
+//
+// Generated by this command:
+//
+//	mockgen -source slackdump.go -destination clienter_mock_test.go -package slackdump -mock_names clienter=mockClienter,Reporter=mockReporter
+//
 // Package slackdump is a generated GoMock package.
 package slackdump
 
@@ -9,8 +13,8 @@ import (
 	io "io"
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	slack "github.com/slack-go/slack"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // mockClienter is a mock of clienter interface.
@@ -46,24 +50,24 @@ func (m *mockClienter) GetConversationHistoryContext(ctx context.Context, params
 }
 
 // GetConversationHistoryContext indicates an expected call of GetConversationHistoryContext.
-func (mr *mockClienterMockRecorder) GetConversationHistoryContext(ctx, params interface{}) *gomock.Call {
+func (mr *mockClienterMockRecorder) GetConversationHistoryContext(ctx, params any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetConversationHistoryContext", reflect.TypeOf((*mockClienter)(nil).GetConversationHistoryContext), ctx, params)
 }
 
 // GetConversationInfoContext mocks base method.
-func (m *mockClienter) GetConversationInfoContext(ctx context.Context, channelID string, includeLocale bool) (*slack.Channel, error) {
+func (m *mockClienter) GetConversationInfoContext(ctx context.Context, input *slack.GetConversationInfoInput) (*slack.Channel, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "GetConversationInfoContext", ctx, channelID, includeLocale)
+	ret := m.ctrl.Call(m, "GetConversationInfoContext", ctx, input)
 	ret0, _ := ret[0].(*slack.Channel)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // GetConversationInfoContext indicates an expected call of GetConversationInfoContext.
-func (mr *mockClienterMockRecorder) GetConversationInfoContext(ctx, channelID, includeLocale interface{}) *gomock.Call {
+func (mr *mockClienterMockRecorder) GetConversationInfoContext(ctx, input any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetConversationInfoContext", reflect.TypeOf((*mockClienter)(nil).GetConversationInfoContext), ctx, channelID, includeLocale)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetConversationInfoContext", reflect.TypeOf((*mockClienter)(nil).GetConversationInfoContext), ctx, input)
 }
 
 // GetConversationRepliesContext mocks base method.
@@ -78,7 +82,7 @@ func (m *mockClienter) GetConversationRepliesContext(ctx context.Context, params
 }
 
 // GetConversationRepliesContext indicates an expected call of GetConversationRepliesContext.
-func (mr *mockClienterMockRecorder) GetConversationRepliesContext(ctx, params interface{}) *gomock.Call {
+func (mr *mockClienterMockRecorder) GetConversationRepliesContext(ctx, params any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetConversationRepliesContext", reflect.TypeOf((*mockClienter)(nil).GetConversationRepliesContext), ctx, params)
 }
@@ -94,9 +98,24 @@ func (m *mockClienter) GetConversationsContext(ctx context.Context, params *slac
 }
 
 // GetConversationsContext indicates an expected call of GetConversationsContext.
-func (mr *mockClienterMockRecorder) GetConversationsContext(ctx, params interface{}) *gomock.Call {
+func (mr *mockClienterMockRecorder) GetConversationsContext(ctx, params any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetConversationsContext", reflect.TypeOf((*mockClienter)(nil).GetConversationsContext), ctx, params)
+}
+
+// GetEmojiContext mocks base method.
+func (m *mockClienter) GetEmojiContext(ctx context.Context) (map[string]string, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetEmojiContext", ctx)
+	ret0, _ := ret[0].(map[string]string)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetEmojiContext indicates an expected call of GetEmojiContext.
+func (mr *mockClienterMockRecorder) GetEmojiContext(ctx any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetEmojiContext", reflect.TypeOf((*mockClienter)(nil).GetEmojiContext), ctx)
 }
 
 // GetFile mocks base method.
@@ -108,7 +127,7 @@ func (m *mockClienter) GetFile(downloadURL string, writer io.Writer) error {
 }
 
 // GetFile indicates an expected call of GetFile.
-func (mr *mockClienterMockRecorder) GetFile(downloadURL, writer interface{}) *gomock.Call {
+func (mr *mockClienterMockRecorder) GetFile(downloadURL, writer any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetFile", reflect.TypeOf((*mockClienter)(nil).GetFile), downloadURL, writer)
 }
@@ -131,7 +150,7 @@ func (mr *mockClienterMockRecorder) GetTeamInfo() *gomock.Call {
 // GetUsersContext mocks base method.
 func (m *mockClienter) GetUsersContext(ctx context.Context, options ...slack.GetUsersOption) ([]slack.User, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{ctx}
+	varargs := []any{ctx}
 	for _, a := range options {
 		varargs = append(varargs, a)
 	}
@@ -142,8 +161,24 @@ func (m *mockClienter) GetUsersContext(ctx context.Context, options ...slack.Get
 }
 
 // GetUsersContext indicates an expected call of GetUsersContext.
-func (mr *mockClienterMockRecorder) GetUsersContext(ctx interface{}, options ...interface{}) *gomock.Call {
+func (mr *mockClienterMockRecorder) GetUsersContext(ctx any, options ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{ctx}, options...)
+	varargs := append([]any{ctx}, options...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetUsersContext", reflect.TypeOf((*mockClienter)(nil).GetUsersContext), varargs...)
+}
+
+// GetUsersInConversationContext mocks base method.
+func (m *mockClienter) GetUsersInConversationContext(ctx context.Context, params *slack.GetUsersInConversationParameters) ([]string, string, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetUsersInConversationContext", ctx, params)
+	ret0, _ := ret[0].([]string)
+	ret1, _ := ret[1].(string)
+	ret2, _ := ret[2].(error)
+	return ret0, ret1, ret2
+}
+
+// GetUsersInConversationContext indicates an expected call of GetUsersInConversationContext.
+func (mr *mockClienterMockRecorder) GetUsersInConversationContext(ctx, params any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetUsersInConversationContext", reflect.TypeOf((*mockClienter)(nil).GetUsersInConversationContext), ctx, params)
 }
