@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: auth.go
-
+//
+// Generated by this command:
+//
+//	mockgen -source=auth.go -destination=../mocks/mock_app/mock_app.go Credentials,createOpener
+//
 // Package mock_app is a generated GoMock package.
 package mock_app
 
@@ -9,8 +13,9 @@ import (
 	io "io"
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
 	auth "github.com/rusq/slackdump/v2/auth"
+	browser "github.com/rusq/slackdump/v2/auth/browser"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockCredentials is a mock of Credentials interface.
@@ -37,18 +42,18 @@ func (m *MockCredentials) EXPECT() *MockCredentialsMockRecorder {
 }
 
 // AuthProvider mocks base method.
-func (m *MockCredentials) AuthProvider(ctx context.Context, workspace string) (auth.Provider, error) {
+func (m *MockCredentials) AuthProvider(ctx context.Context, workspace string, browser browser.Browser, legacy bool) (auth.Provider, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "AuthProvider", ctx, workspace)
+	ret := m.ctrl.Call(m, "AuthProvider", ctx, workspace, browser, legacy)
 	ret0, _ := ret[0].(auth.Provider)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
 // AuthProvider indicates an expected call of AuthProvider.
-func (mr *MockCredentialsMockRecorder) AuthProvider(ctx, workspace interface{}) *gomock.Call {
+func (mr *MockCredentialsMockRecorder) AuthProvider(ctx, workspace, browser, legacy any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AuthProvider", reflect.TypeOf((*MockCredentials)(nil).AuthProvider), ctx, workspace)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AuthProvider", reflect.TypeOf((*MockCredentials)(nil).AuthProvider), ctx, workspace, browser, legacy)
 }
 
 // IsEmpty mocks base method.
@@ -98,7 +103,7 @@ func (m *MockcreateOpener) Create(arg0 string) (io.WriteCloser, error) {
 }
 
 // Create indicates an expected call of Create.
-func (mr *MockcreateOpenerMockRecorder) Create(arg0 interface{}) *gomock.Call {
+func (mr *MockcreateOpenerMockRecorder) Create(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Create", reflect.TypeOf((*MockcreateOpener)(nil).Create), arg0)
 }
@@ -113,7 +118,7 @@ func (m *MockcreateOpener) Open(arg0 string) (io.ReadCloser, error) {
 }
 
 // Open indicates an expected call of Open.
-func (mr *MockcreateOpenerMockRecorder) Open(arg0 interface{}) *gomock.Call {
+func (mr *MockcreateOpenerMockRecorder) Open(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Open", reflect.TypeOf((*MockcreateOpener)(nil).Open), arg0)
 }
