@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: github.com/rusq/slackdump/v2/downloader (interfaces: Downloader)
-
+//
+// Generated by this command:
+//
+//	mockgen -destination internal/mocks/mock_downloader/mock_downloader.go github.com/rusq/slackdump/v2/downloader Downloader
+//
 // Package mock_downloader is a generated GoMock package.
 package mock_downloader
 
@@ -8,7 +12,7 @@ import (
 	io "io"
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockDownloader is a mock of Downloader interface.
@@ -43,7 +47,7 @@ func (m *MockDownloader) GetFile(arg0 string, arg1 io.Writer) error {
 }
 
 // GetFile indicates an expected call of GetFile.
-func (mr *MockDownloaderMockRecorder) GetFile(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockDownloaderMockRecorder) GetFile(arg0, arg1 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetFile", reflect.TypeOf((*MockDownloader)(nil).GetFile), arg0, arg1)
 }
