@@ -1,13 +1,17 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: io (interfaces: ReadCloser,WriteCloser)
-
+//
+// Generated by this command:
+//
+//	mockgen -destination=../mocks/mock_io/mock_io.go io ReadCloser,WriteCloser
+//
 // Package mock_io is a generated GoMock package.
 package mock_io
 
 import (
 	reflect "reflect"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockReadCloser is a mock of ReadCloser interface.
@@ -57,7 +61,7 @@ func (m *MockReadCloser) Read(arg0 []byte) (int, error) {
 }
 
 // Read indicates an expected call of Read.
-func (mr *MockReadCloserMockRecorder) Read(arg0 interface{}) *gomock.Call {
+func (mr *MockReadCloserMockRecorder) Read(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Read", reflect.TypeOf((*MockReadCloser)(nil).Read), arg0)
 }
@@ -109,7 +113,7 @@ func (m *MockWriteCloser) Write(arg0 []byte) (int, error) {
 }
 
 // Write indicates an expected call of Write.
-func (mr *MockWriteCloserMockRecorder) Write(arg0 interface{}) *gomock.Call {
+func (mr *MockWriteCloserMockRecorder) Write(arg0 any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Write", reflect.TypeOf((*MockWriteCloser)(nil).Write), arg0)
 }
