@@ -1,6 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: os (interfaces: FileInfo)
-
+//
+// Generated by this command:
+//
+//	mockgen -destination internal/mocks/mock_os/mock_os.go os FileInfo
+//
 // Package mock_os is a generated GoMock package.
 package mock_os
 
@@ -9,7 +13,7 @@ import (
 	reflect "reflect"
 	time "time"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockFileInfo is a mock of FileInfo interface.
@@ -106,10 +110,10 @@ func (mr *MockFileInfoMockRecorder) Size() *gomock.Call {
 }
 
 // Sys mocks base method.
-func (m *MockFileInfo) Sys() interface{} {
+func (m *MockFileInfo) Sys() any {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Sys")
-	ret0, _ := ret[0].(interface{})
+	ret0, _ := ret[0].(any)
 	return ret0
 }
 
