@@ -8,8 +8,10 @@ import (
 	"strings"
 
 	"github.com/fatih/color"
+	"golang.org/x/term"
 )
 
+// CLI is the archaic fallback UI for auth.
 type CLI struct{}
 
 func (*CLI) instructions(w io.Writer) {
@@ -24,22 +26,92 @@ func (*CLI) instructions(w io.Writer) {
 
 func (cl *CLI) RequestWorkspace(w io.Writer) (string, error) {
 	cl.instructions(w)
-	fmt.Fprint(w, "Enter Slack Workspace Name: ")
-	workspace, err := readln(os.Stdin)
+	workspace, err := prompt(w, "Enter Slack Workspace Name or URL: ", readln)
 	if err != nil {
 		return "", err
 	}
-	return workspace, nil
+	return Sanitize(workspace)
 }
 
-func (*CLI) Stop() {
-	return
+func (cl *CLI) RequestEmail(w io.Writer) (string, error) {
+	return prompt(w, "Enter Email: ", readln)
 }
 
-func readln(r io.Reader) (string, error) {
+func (cl *CLI) RequestPassword(w io.Writer, account string) (string, error) {
+	defer fmt.Fprintln(w)
+	return prompt(w, fmt.Sprintf("Enter Password for %s (won't be visible): ", account), readpwd)
+}
+
+func (cl *CLI) RequestLoginType(w io.Writer) (LoginType, error) {
+	var types = []struct {
+		name  string
+		value LoginType
+	}{
+		{"Email", LHeadless},
+		{"Google", LInteractive},
+		{"Apple", LInteractive},
+		{"Login with Single-Sign-On (SSO)", LInteractive},
+		{"Other/Manual", LInteractive},
+		{"Cancel", LCancel},
+	}
+
+	var idx int = -1
+	for idx < 0 || idx >= len(types) {
+		fmt.Fprintf(w, "Select login type:\n")
+		for i, t := range types {
+			fmt.Fprintf(w, "\t%d. %s\n", i+1, t.name)
+		}
+		fmt.Fprintf(w, "Enter number, and press Enter: ")
+
+		_, err := fmt.Fscanf(os.Stdin, "%d", &idx)
+		if err != nil {
+			fmt.Fprintln(w, err)
+			continue
+		}
+
+		idx -= 1 // adjusting for 0-index
+
+		if idx < 0 || idx >= len(types) {
+			fmt.Fprintln(w, "invalid login type")
+		}
+	}
+	return types[idx].value, nil
+}
+
+func (*CLI) Stop() {}
+
+func readln(r *os.File) (string, error) {
 	line, err := bufio.NewReader(r).ReadString('\n')
 	if err != nil {
 		return "", err
 	}
 	return strings.TrimSpace(line), nil
+}
+
+func readpwd(f *os.File) (string, error) {
+	pwd, err := term.ReadPassword(int(f.Fd()))
+	if err != nil {
+		return "", err
+	}
+	return string(pwd), nil
+}
+
+func prompt(w io.Writer, prompt string, readlnFn func(*os.File) (string, error)) (string, error) {
+	for {
+		fmt.Fprint(w, prompt)
+		v, err := readlnFn(os.Stdin)
+		if err != nil {
+			return "", err
+		}
+		if v != "" {
+			return v, nil
+		}
+		fmt.Fprintln(w, "input cannot be empty")
+	}
+}
+
+func (*CLI) ConfirmationCode(email string) (code int, err error) {
+	fmt.Printf("Enter confirmation code sent to %s: ", email)
+	_, err = fmt.Fscanf(os.Stdin, "%d", &code)
+	return
 }
