@@ -22,11 +22,10 @@ import (
 )
 
 const (
-	defDownloadDir = "."  // default download directory is current.
-	defRetries     = 3    // default number of retries if download fails
-	defNumWorkers  = 4    // number of download processes
-	defLimit       = 5000 // default API limit, in events per second.
-	defFileBufSz   = 100  // default download channel buffer.
+	defRetries    = 3    // default number of retries if download fails
+	defNumWorkers = 4    // number of download processes
+	defLimit      = 5000 // default API limit, in events per second.
+	defFileBufSz  = 100  // default download channel buffer.
 )
 
 // Client is the instance of the downloader.
@@ -40,10 +39,19 @@ type Client struct {
 	workers int
 
 	mu           sync.Mutex // mutex prevents race condition when starting/stopping
-	fileRequests chan FileRequest
+	fileRequests chan fileRequest
 	wg           *sync.WaitGroup
 	started      bool
+
+	nameFn FilenameFunc
 }
+
+// FilenameFunc is the file naming function that should return the output
+// filename for slack.File.
+type FilenameFunc func(*slack.File) string
+
+// Filename returns name of the file generated from the slack.File.
+var Filename FilenameFunc = stdFilenameFn
 
 // Downloader is the file downloader interface.  It exists primarily for mocking
 // in tests.
@@ -95,6 +103,16 @@ func Logger(l logger.Interface) Option {
 	}
 }
 
+func WithNameFunc(fn FilenameFunc) Option {
+	return func(c *Client) {
+		if fn != nil {
+			c.nameFn = fn
+		} else {
+			c.nameFn = Filename
+		}
+	}
+}
+
 // New initialises new file downloader.
 func New(client Downloader, fs fsadapter.FS, opts ...Option) *Client {
 	if client == nil {
@@ -107,6 +125,7 @@ func New(client Downloader, fs fsadapter.FS, opts ...Option) *Client {
 		limiter: rate.NewLimiter(defLimit, 1),
 		retries: defRetries,
 		workers: defNumWorkers,
+		nameFn:  Filename,
 	}
 	for _, opt := range opts {
 		opt(c)
@@ -119,13 +138,13 @@ func (c *Client) SaveFile(ctx context.Context, dir string, f *slack.File) (int64
 	return c.saveFile(ctx, dir, f)
 }
 
-type FileRequest struct {
+type fileRequest struct {
 	Directory string
 	File      *slack.File
 }
 
-// Start starts an async file downloader.  If the downloader
-// is already started, it does nothing.
+// Start starts an async file downloader.  If the downloader is already
+// started, it does nothing.
 func (c *Client) Start(ctx context.Context) {
 	c.mu.Lock()
 	defer c.mu.Unlock()
@@ -134,7 +153,7 @@ func (c *Client) Start(ctx context.Context) {
 		// already started
 		return
 	}
-	req := make(chan FileRequest, defFileBufSz)
+	req := make(chan fileRequest, defFileBufSz)
 
 	c.fileRequests = req
 	c.wg = c.startWorkers(ctx, req)
@@ -143,16 +162,17 @@ func (c *Client) Start(ctx context.Context) {
 
 // startWorkers starts download workers.  It returns a sync.WaitGroup.  If the
 // req channel is closed, workers will stop, and wg.Wait() completes.
-func (c *Client) startWorkers(ctx context.Context, req <-chan FileRequest) *sync.WaitGroup {
+func (c *Client) startWorkers(ctx context.Context, req <-chan fileRequest) *sync.WaitGroup {
 	if c.workers == 0 {
 		c.workers = defNumWorkers
 	}
+	seenC := c.fltSeen(req)
 	var wg sync.WaitGroup
 	// create workers
 	for i := 0; i < c.workers; i++ {
 		wg.Add(1)
 		go func(workerNum int) {
-			c.worker(ctx, c.fltSeen(req))
+			c.worker(ctx, seenC)
 			wg.Done()
 			c.l().Debugf("download worker %d terminated", workerNum)
 		}(i)
@@ -162,7 +182,7 @@ func (c *Client) startWorkers(ctx context.Context, req <-chan FileRequest) *sync
 
 // worker receives requests from reqC and passes them to saveFile function.
 // It will stop if either context is Done, or reqC is closed.
-func (c *Client) worker(ctx context.Context, reqC <-chan FileRequest) {
+func (c *Client) worker(ctx context.Context, reqC <-chan fileRequest) {
 	for {
 		select {
 		case <-ctx.Done():
@@ -172,18 +192,18 @@ func (c *Client) worker(ctx context.Context, reqC <-chan FileRequest) {
 			if !moar {
 				return
 			}
-			c.l().Debugf("saving %q to %s, size: %d", Filename(req.File), req.Directory, req.File.Size)
+			c.l().Debugf("saving %q to %s, size: %d", c.nameFn(req.File), req.Directory, req.File.Size)
 			n, err := c.saveFile(ctx, req.Directory, req.File)
 			if err != nil {
-				c.l().Printf("error saving %q to %q: %s", Filename(req.File), req.Directory, err)
+				c.l().Printf("error saving %q to %q: %s", c.nameFn(req.File), req.Directory, err)
 				break
 			}
-			c.l().Printf("file %q saved to %s: %d bytes written", Filename(req.File), req.Directory, n)
+			c.l().Printf("file %q saved to %s: %d bytes written", c.nameFn(req.File), req.Directory, n)
 		}
 	}
 }
 
-var errNoFS = errors.New("fs adapter not initialised")
+var ErrNoFS = errors.New("fs adapter not initialised")
 
 // AsyncDownloader starts Client.worker goroutines to download files
 // concurrently. It will download any file that is received on fileDlQueue
@@ -191,15 +211,15 @@ var errNoFS = errors.New("fs adapter not initialised")
 // closed once all downloads are complete.
 func (c *Client) AsyncDownloader(ctx context.Context, dir string, fileDlQueue <-chan *slack.File) (chan struct{}, error) {
 	if c.fs == nil {
-		return nil, errNoFS
+		return nil, ErrNoFS
 	}
 	done := make(chan struct{})
 
-	req := make(chan FileRequest)
+	req := make(chan fileRequest)
 	go func() {
 		defer close(req)
 		for f := range fileDlQueue {
-			req <- FileRequest{Directory: dir, File: f}
+			req <- fileRequest{Directory: dir, File: f}
 		}
 	}()
 
@@ -217,9 +237,13 @@ func (c *Client) AsyncDownloader(ctx context.Context, dir string, fileDlQueue <-
 // saveFileWithLimiter saves the file to specified directory, it will use the provided limiter l for throttling.
 func (c *Client) saveFile(ctx context.Context, dir string, sf *slack.File) (int64, error) {
 	if c.fs == nil {
-		return 0, errNoFS
+		return 0, ErrNoFS
 	}
-	filePath := filepath.Join(dir, Filename(sf))
+	if mode := sf.Mode; mode == "hidden_by_limit" || mode == "external" || sf.IsExternal {
+		trace.Logf(ctx, "info", "file %q is not downloadable", sf.Name)
+		return 0, nil
+	}
+	filePath := filepath.Join(dir, c.nameFn(sf))
 
 	tf, err := os.CreateTemp("", "")
 	if err != nil {
@@ -265,8 +289,7 @@ func (c *Client) saveFile(ctx context.Context, dir string, sf *slack.File) (int6
 	return int64(n), nil
 }
 
-// Filename returns name of the file
-func Filename(f *slack.File) string {
+func stdFilenameFn(f *slack.File) string {
 	return fmt.Sprintf("%s-%s", f.ID, f.Name)
 }
 
@@ -304,7 +327,7 @@ func (c *Client) DownloadFile(dir string, f slack.File) (string, error) {
 	if !started {
 		return "", ErrNotStarted
 	}
-	c.fileRequests <- FileRequest{Directory: dir, File: &f}
+	c.fileRequests <- fileRequest{Directory: dir, File: &f}
 	return path.Join(dir, Filename(&f)), nil
 }
 
