@@ -1,15 +1,21 @@
 package export
 
 import (
+	"context"
+	"encoding/json"
+	"os"
+	"runtime/trace"
 	"testing"
+	"time"
 
 	"github.com/rusq/slackdump/v2/internal/fixtures"
+	"github.com/rusq/slackdump/v2/internal/fixtures/fixgen"
 	"github.com/rusq/slackdump/v2/types"
 	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestConversation_ByDate(t *testing.T) {
-	// TODO
 	var exp Export
 
 	conversations := fixtures.Load[types.Conversation](fixtures.TestConversationJSON)
@@ -21,10 +27,33 @@ func TestConversation_ByDate(t *testing.T) {
 	}
 
 	// uncomment to write the json for fixtures
-	// require.NoError(t, writeOutput("convDt", convDt))
+	require.NoError(t, writeOutput("convDt", convDt))
 
-	want := fixtures.Load[map[string][]ExportMessage](fixtures.TestConversationExportJSON)
+	want := fixtures.Load[messagesByDate](fixtures.TestConversationExportJSON)
+
+	// we need to depopulate slackdumpTime for comparison, as it is not saved
+	// in the fixture.
+	zeroSlackdumpTime(convDt)
 	assert.Equal(t, want, convDt)
+}
+
+func zeroSlackdumpTime(m messagesByDate) {
+	for _, msgs := range m {
+		for i := range msgs {
+			msgs[i].slackdumpTime = time.Time{}
+		}
+	}
+}
+
+func writeOutput(name string, v interface{}) error {
+	f, err := os.Create(name + ".json")
+	if err != nil {
+		return err
+	}
+	defer f.Close()
+	enc := json.NewEncoder(f)
+	enc.SetIndent("", "\t")
+	return enc.Encode(v)
 }
 
 func Test_messagesByDate_validate(t *testing.T) {
@@ -35,22 +64,22 @@ func Test_messagesByDate_validate(t *testing.T) {
 	}{
 		{"valid",
 			messagesByDate{
-				"2019-09-16": []ExportMessage{},
-				"2020-12-31": []ExportMessage{},
+				"2019-09-16": []*ExportMessage{},
+				"2020-12-31": []*ExportMessage{},
 			},
 			false,
 		},
 		{"empty key",
 			messagesByDate{
-				"":           []ExportMessage{},
-				"2020-12-31": []ExportMessage{},
+				"":           []*ExportMessage{},
+				"2020-12-31": []*ExportMessage{},
 			},
 			true,
 		},
 		{"invalid key",
 			messagesByDate{
-				"2019-09-16": []ExportMessage{},
-				"2020-31-12": []ExportMessage{}, //swapped month and date
+				"2019-09-16": []*ExportMessage{},
+				"2020-31-12": []*ExportMessage{}, //swapped month and date
 			},
 			true,
 		},
@@ -62,4 +91,39 @@ func Test_messagesByDate_validate(t *testing.T) {
 			}
 		})
 	}
+}
+
+var (
+	benchResult messagesByDate
+	benchConv   types.Conversation
+)
+
+func init() {
+	var (
+		startDate   = time.Date(2020, 1, 1, 0, 0, 0, 0, time.UTC)
+		endDate     = time.Date(2020, 1, 1, 15, 0, 0, 0, time.UTC)
+		numMessages = 10_000
+	)
+	benchConv = fixgen.GenerateTestConversation("test", startDate, endDate, numMessages)
+}
+
+func BenchmarkByDate(b *testing.B) {
+
+	ctx, task := trace.NewTask(context.Background(), "BenchmarkByDate")
+	defer task.End()
+
+	var (
+		ex  Export
+		err error
+	)
+	region := trace.StartRegion(ctx, "byDateBenchRun")
+	defer region.End()
+	var m messagesByDate
+	for i := 0; i < b.N; i++ {
+		m, err = ex.byDate(&benchConv, nil)
+		if err != nil {
+			b.Fatal(err)
+		}
+	}
+	benchResult = m
 }
