@@ -9,9 +9,14 @@ import (
 
 // Options allows to configure slack export options.
 type Options struct {
-	Oldest       time.Time
-	Latest       time.Time
-	IncludeFiles bool
-	Logger       logger.Interface
-	List         *structures.EntityList
+	Oldest      time.Time
+	Latest      time.Time
+	Logger      logger.Interface
+	List        *structures.EntityList
+	Type        ExportType
+	ExportToken string
+}
+
+func (opt Options) IsFilesEnabled() bool {
+	return opt.Type > TNoDownload
 }
