@@ -1,6 +1,6 @@
 package fixtures
 
-//TestConversationJSON channel: CHM82GF99
+// TestConversationJSON channel: CHM82GF99
 const TestConversationJSON = `
 {
 	"name": "everything",
@@ -324,5 +324,74 @@ const TestConversationJSON = `
 	  }
 	],
 	"channel_id": "CHM82GF99"
+  }
+`
+
+const TestChannel = `
+  {
+    "id": "CLPGTGPAA",
+    "created": 1563771279,
+    "is_open": false,
+    "last_read": "1656044916.988719",
+    "is_group": false,
+    "is_shared": false,
+    "is_im": false,
+    "is_ext_shared": false,
+    "is_org_shared": false,
+    "is_pending_ext_shared": false,
+    "is_private": false,
+    "is_mpim": false,
+    "unlinked": 0,
+    "name_normalized": "random",
+    "num_members": 0,
+    "priority": 0,
+    "user": "",
+    "name": "random",
+    "creator": "LOL1",
+    "is_archived": false,
+    "members": null,
+    "topic": {
+      "value": "Non-work banter and water cooler conversation",
+      "creator": "LOL1",
+      "last_set": 1563771279
+    },
+    "purpose": {
+      "value": "A place for non-work-related flimflam, faffing, hodge-podge or jibber-jabber you'd prefer to keep out of more focused work-related channels.",
+      "creator": "LOL1",
+      "last_set": 1563771279
+    },
+    "is_channel": true,
+    "is_general": false,
+    "is_member": true,
+    "locale": "en-US"
+  }
+`
+
+const TestMessage = `	  {
+	"client_msg_id": "c6cdfb3a-59d6-4198-9800-cc74bcdc0b7d",
+	"type": "message",
+	"user": "UHSD97ZA5",
+	"text": "Test message with Html chars \u0026lt; \u0026gt;",
+	"ts": "1645095505.023899",
+	"team": "THY5HTZ8U",
+	"replace_original": false,
+	"delete_original": false,
+	"blocks": [
+	  {
+		"type": "rich_text",
+		"block_id": "SkX",
+		"elements": [
+		  {
+			"type": "rich_text_section",
+			"elements": [
+			  {
+				"type": "text",
+				"text": "Test message with Html chars \u003c \u003e"
+			  }
+			]
+		  }
+		]
+	  }
+	]
   }
 `
