@@ -20,11 +20,42 @@ func NewUserIndex(us []slack.User) UserIndex {
 	return usermap
 }
 
-// Username tries to resolve the username by ID. If the user index is not
-// initialised, it will return the ID, otherwise, if the user is not found in
-// cache, it will assume that the user is external, and return the ID with
-// "external" prefix.
+// Username tries to resolve the username by ID. If it fails, it returns the
+// user ID.  If the user is not found in index, is assumes that it is an external
+// user and returns ID with "external" prefix.
 func (idx UserIndex) Username(id string) string {
+	return idx.userattr(id, func(user *slack.User) string {
+		return user.Name
+	})
+}
+
+// DisplayName tries to resolve the display name by ID. if the index is empty, it
+// returns the user ID. If the user is not found in index, is assumes that it is
+// an external user and returns ID with "external" prefix. If it does find the
+// user and display name is unavailble, it returns the Real Name.
+func (idx UserIndex) DisplayName(id string) string {
+	return idx.userattr(id, func(user *slack.User) string {
+		return nvl(user.Profile.DisplayName, user.RealName)
+	})
+}
+
+func nvl(s string, ss ...string) string {
+	if s != "" {
+		return s
+	}
+	for _, alt := range ss {
+		if alt != "" {
+			return alt
+		}
+	}
+	return "" // you got no luck at all, don't you.
+}
+
+// userattr finds the user by ID and calls a function fn with that user. If the
+// user index is not initialised, it will return the ID, otherwise, if the user
+// is not found in index, it will assume that the user is external, and return
+// the ID with "external" prefix.
+func (idx UserIndex) userattr(id string, fn func(user *slack.User) string) string {
 	if idx == nil {
 		// no user cache, use the IDs.
 		return id
@@ -33,7 +64,7 @@ func (idx UserIndex) Username(id string) string {
 	if !ok {
 		return "<external>:" + id
 	}
-	return user.Name
+	return fn(user)
 }
 
 // Sender returns username for the message
@@ -46,7 +77,7 @@ func (idx UserIndex) Sender(msg *slack.Message) string {
 	}
 
 	if userid != "" {
-		return idx.Username(userid)
+		return idx.DisplayName(userid)
 	}
 
 	return ""
