@@ -7,9 +7,9 @@ import (
 
 	"errors"
 
-	"github.com/golang/mock/gomock"
 	"github.com/slack-go/slack"
 	"github.com/stretchr/testify/assert"
+	"go.uber.org/mock/gomock"
 	"golang.org/x/time/rate"
 
 	"github.com/rusq/slackdump/v2/internal/network"
