@@ -6,6 +6,7 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
+	"io"
 	"os"
 	"path/filepath"
 	"runtime/trace"
@@ -55,7 +56,7 @@ func (sd *Session) fetchUsers(ctx context.Context) (types.Users, error) {
 	var (
 		users []slack.User
 	)
-	if err := withRetry(ctx, network.NewLimiter(network.Tier2, sd.options.Tier2Burst, int(sd.options.Tier2Boost)), sd.options.Tier2Retries, func() error {
+	if err := network.WithRetry(ctx, network.NewLimiter(network.Tier2, sd.options.Tier2Burst, int(sd.options.Tier2Boost)), sd.options.Tier2Retries, func() error {
 		var err error
 		users, err = sd.client.GetUsersContext(ctx)
 		return err
@@ -86,10 +87,26 @@ func (sd *Session) loadUserCache(filename string, suffix string, maxAge time.Dur
 	}
 	defer f.Close()
 
-	dec := json.NewDecoder(f)
-	var uu types.Users
-	if err := dec.Decode(&uu); err != nil {
+	uu, err := readUsers(f)
+	if err != nil {
 		return nil, fmt.Errorf("failed to decode users from %s: %w", filename, err)
+	}
+
+	return uu, nil
+}
+
+func readUsers(r io.Reader) (types.Users, error) {
+	dec := json.NewDecoder(r)
+	var uu = make(types.Users, 0, 500) // 500 users. reasonable?
+	for {
+		var u slack.User
+		if err := dec.Decode(&u); err != nil {
+			if err == io.EOF {
+				break
+			}
+			return nil, err
+		}
+		uu = append(uu, u)
 	}
 	return uu, nil
 }
@@ -104,8 +121,10 @@ func (sd *Session) saveUserCache(filename string, suffix string, uu types.Users)
 	defer f.Close()
 
 	enc := json.NewEncoder(f)
-	if err := enc.Encode(uu); err != nil {
-		return fmt.Errorf("failed to encode data for %s: %w", filename, err)
+	for _, u := range uu {
+		if err := enc.Encode(u); err != nil {
+			return fmt.Errorf("failed to encode data for %s: %w", filename, err)
+		}
 	}
 	return nil
 }
